Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -3,7 +3,7 @@
 
 int main() {
     float rad;
-    int pi = 3.14;
+    float pi = 3.14;
     scanf("%f",&rad);
     printf("Area: %.2f\n",pi*rad*rad);
     return 0;
